neopixel: parameterize by color layout
This allows abstracting the number and order of neopixel color
channels (e.g. RGBW, GRB, etc). To achieve this, we start extending
the RegValue type to behave more like a buffer with indexing traits
and in the process move it and Reg into their own reg module. It is
also now mandatory to supply a color layout when instantiating a
Neopixel module.