keypad: remove x/y indexing and event initialization
The dimensions of the keypad are unknown and users will almost
certainly have their own ideas about a sensible indexing scheme so
simply use key indexes instead of trying to translate to a cartesion
coordinate scheme. Also, enabling events is rather slow and isn't
one-size-fits-all so remove it from initialization and leave it to the
user. This makes it easier to do things like check for held keys
during POST to enter configuration modes and it may also interfere
with interrupt enablement (TBD).

Also, run cargo fmt and cargo clippy.